Remove old figures and update SK animation to not remake files unnecessary
Former-commit-id: 48c6332086f43afbfedd313e0a74944b5376ff24
Former-commit-id: 39398cc5940056cbe6157b2e60624152c63c902c
Former-commit-id: 1aba077a2423032a72adae120caf72fdc508b20d
Former-commit-id: 537825dac7f787a49b5e9de236f694d4180a92aa
Former-commit-id: e458f9ef4f7c099fd23f54314d18712b92efb6c1 [formerly a3558cd36f6cc6894e5f5360df4e342b1a36cffe]
Former-commit-id: 0456e9b46c168565d7350534c34743595c445bfa

+++ SK_P1_UK_analysis.py
@@ -234,7 +234,7 @@ class AnimateSIR:
         if not Path(gifname).exists() or force_rerun:
             if self.verbose and not self.do_tqdm:
                 print("\nMake individual frames", flush=True)
-            self._make_png_files(dpi, self.do_tqdm)
+            self._make_png_files(dpi, self.do_tqdm, force_rerun)
 
             if self.verbose:
                 print("\nMake GIF", flush=True)
@@ -259,17 +259,18 @@ class AnimateSIR:
         sim_pars_str = self._get_sim_pars_str()
         return f"Figures_SK_P1_UK/tmp_{sim_pars_str}/animation_{sim_pars_str}_frame_{i_day:06d}.png"
 
-    def _make_png_files(self, dpi, do_tqdm):
+    def _make_png_files(self, dpi, do_tqdm, force_rerun=False):
         it = range(self.N)
         if do_tqdm:
             it = tqdm(it, desc='Make individual frames')
         for i_day in it:
-            fig, _ = self._plot_df_i_day(i_day, dpi)
             png_name = self._get_png_name(i_day)
-            Path(png_name).parent.mkdir(parents=True, exist_ok=True)
-            fig.savefig(png_name, dpi=dpi, bbox_inches='tight', pad_inches=0.3)
-            plt.close(fig)
-            plt.close('all')
+            if not Path(png_name).exists() or force_rerun:
+                fig, _ = self._plot_df_i_day(i_day, dpi)
+                Path(png_name).parent.mkdir(parents=True, exist_ok=True)
+                fig.savefig(png_name, dpi=dpi, bbox_inches='tight', pad_inches=0.3)
+                plt.close(fig)
+                plt.close('all')
         return None
 
     def _make_gif_file(self, gifname):
